finans_backend/requirements.txt
django
waitress
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
django-celery-beat
yfinance
requests
